HelpEmbedFactory and better fastSend
Created a new function in MessageUtils called HelpEmbedFactory to
generate a embed with the command information in only one line.
FastSend now accepts a third argument, you can give the variable context
to the i18n using it, it's the same way, Send(string, boolean, object),
the value of the boolean should be false so it uses the object.
Fixed some documentations typos.